src/game.ts
export class Hex {
  row: number;
  col: number;
  isEmpty: boolean;
  isInaccessible: boolean;
  player: Player | null;
  forces: number;

  constructor(row: number, col: number) {
    this.row = row;
    this.col = col;
    this.isEmpty = true;
    this.isInaccessible = false;
    this.player = null;
    this.forces = 0;
  }
}

export class Player {
  id: number;
  color: number;
  isAI: boolean;
  forces: number;
  diplomaticStatus: Map<number, 'Peace' | 'War'>;

  constructor(id: number, color: number, isAI: boolean) {
    this.id = id;
    this.color = color;
    this.isAI = isAI;
    this.forces = 0;
    this.diplomaticStatus = new Map();
  }

  distributeForces(hexes: Hex[]) {
    // Placeholder for distributing forces logic
  }

  moveUnits(hex: Hex, targetHex: Hex) {
    // Placeholder for moving units logic
  }

  attack(hex: Hex, targetHex: Hex) {
    // Placeholder for attack logic
  }

  chooseDiplomaticStatus(otherPlayer: Player, status: 'Peace' | 'War') {
    this.diplomaticStatus.set(otherPlayer.id, status);
  }
}

export class Game {
  mapSize: number;
  hexes: Hex[][];
  players: Player[];
  currentPlayerIndex: number;

  constructor(mapSize: number, players: Player[]) {
    this.mapSize = mapSize;
    this.hexes = [];
    this.players = players;
    this.currentPlayerIndex = 0;

    for (let row = 0; row < mapSize; row++) {
      this.hexes[row] = [];
      for (let col = 0; col < mapSize; col++) {
        this.hexes[row][col] = new Hex(row, col);
      }
    }

    this.generateMap();
  }

  generateMap() {
    // Placeholder for map generation logic
  }

  getHex(row: number, col: number): Hex | null {
    if (row >= 0 && row < this.mapSize && col >= 0 && col < this.mapSize) {
      return this.hexes[row][col];
    }
    return null;
  }

  start() {
    // Placeholder for game start logic
  }

  update() {
    // Placeholder for game update logic
  }

  onStateChange(callback: () => void) {
    // Placeholder for state change callback logic
  }
}
